backend/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
sqlparse==0.5.0
requests==2.26.0
django-allauth>=0.54.0
dj-rest-auth>=2.2.5
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0